adelanto - aguinaldos sin terminar

+++ BusinessLogic/Pago.cs
@@ -14,15 +14,15 @@ namespace BusinessLogic
         public Int32 Id { get; set; }
         public Int32 IdColaborador { get; set; }
         public DateTime FechaPago { get; set; }
-        public Double SalarioBruto { get; set; }
-        public Double SalarioNeto { get; set; }
-        public Double Rebajo { get; set; }
-        public Double HorasLaboradas { get; set; }
-        public Double HorasExtra { get; set; }
+        public Decimal SalarioBruto { get; set; }
+        public Decimal SalarioNeto { get; set; }
+        public Decimal Rebajo { get; set; }
+        public Decimal HorasLaboradas { get; set; }
+        public Decimal HorasExtra { get; set; }
         public String TransferenciaPago { get; set; }
         public Boolean EstadoPago { get; set; }
 
-        public Pago(Int32 id, Int32 idColaborador, DateTime fecha, Double salarioBruto, Double salarioNeto, Double rebajo, Double horasLaboradas, Double horasExtra, String transferencia, Boolean estado)
+        public Pago(Int32 id, Int32 idColaborador, DateTime fecha, Decimal salarioBruto, Decimal salarioNeto, Decimal rebajo, Decimal horasLaboradas, Decimal horasExtra, String transferencia, Boolean estado)
         {
             this.Id = id;
             this.IdColaborador = idColaborador;
